src/data-ingestion/requirement.txt
requests
mysql-connector-python
datetime